Update solana.tokenlist.json
NaruToken (NRT) is a fun, community-driven memecoin inspired by the beloved Naruto universe. Built on the fast and scalable Solana blockchain, NaruToken is designed to bring fans together while providing exciting opportunities for engagement and growth. With a total supply of 5 billion tokens, NaruToken focuses on fostering a vibrant community and delivering endless fun for its holders. Join the ninja revolution and be part of the journey with NRT!